Use long for expires

+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/BanFromCommunity.kt
@@ -9,6 +9,6 @@ internal data class BanFromCommunity(
     val ban: Boolean,
     val remove_data: Boolean? = null,
     val reason: String? = null,
-    val expires: Int? = null,
+    val expires: Long? = null,
     val auth: String,
 )
